refactor: LottoConstants 지역화
