--- src/main/java/lotto/constants/LottoConstants.java
@@ -1,16 +1,0 @@
-package lotto.constants;
-
-public class LottoConstants {
-    private LottoConstants() {
-    }
-
-    public static final int MIN_LOTTO_NUM = 1;
-    public static final int MAX_LOTTO_NUM = 45;
-    public static final int MAX_LOTTO_NUMBER_SIZE = 6;
-
-    public static final int LOTTO_TICKET_PRICE = 1_000;
-
-    public static final String NUMBERS_DELIMITER = ",";
-    public static final String NUMBERS_PREFIX = "[";
-    public static final String NUMBERS_SUFFIX = "]";
-}
