+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,6 @@
 package lotto.view;
 
-import static lotto.constants.LottoConstants.NUMBERS_DELIMITER;
+import static lotto.domain.LottoNumbers.NUMBERS_DELIMITER;
 
 import java.util.Arrays;
 import java.util.List;
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 import lotto.domain.Money;
 
 public class InputView {
+
     public static final String INPUT_MONEY = "구입금액을 입력해 주세요.";
     public static final String INPUT_WINNING_LOTTO = "지난 주 당첨 번호를 입력해 주세요.";
 
@@ -38,6 +39,5 @@ public class InputView {
     private static String readLine() {
         return scanner.nextLine();
     }
-
 
 }
